cognitive services/requirements.txt
python-dotenv
azure-ai-formrecognizer
azure.core
azure-cognitiveservices-vision-computervision
Pillow